fix: make Dockerode preload fail open
